+++ src/main/scala/wdlTools/cli/Upgrade.scala
@@ -1,11 +1,8 @@
 package wdlTools.cli
 
-import java.nio.file.Files
-
 import wdlTools.generators.code.Upgrader
-import dx.util.{FileSourceResolver, LocalFileSource}
+import dx.util.{FileSourceResolver, FileUtils, LocalFileSource}
 
-import scala.jdk.CollectionConverters._
 import scala.language.reflectiveCalls
 
 case class Upgrade(conf: WdlToolsConf) extends Command {
@@ -20,9 +17,13 @@ case class Upgrade(conf: WdlToolsConf) extends Command {
     val documents = upgrader.upgrade(docSource, srcVersion, destVersion)
     documents.foreach {
       case (source, lines) if outputDir.isDefined =>
-        Files.write(outputDir.get.resolve(source.name), lines.asJava)
+        FileUtils.writeFileContent(outputDir.get.resolve(source.name),
+                                   lines.mkString(System.lineSeparator()),
+                                   overwrite = overwrite)
       case (localSource: LocalFileSource, lines) if overwrite =>
-        Files.write(localSource.canonicalPath, lines.asJava)
+        FileUtils.writeFileContent(localSource.canonicalPath,
+                                   lines.mkString(System.lineSeparator()),
+                                   overwrite = overwrite)
       case (_, lines) =>
         println(lines.mkString(System.lineSeparator()))
     }