+++ RELEASE_NOTES.md
@@ -1,7 +1,13 @@
 # Change log
 
+## 0.14.0 (2021-05-27)
+
+* Adds `fix` tool to automatically fix some non-compliant WDL syntax
+* Improve error messages in expression evaluator
+
 ## 0.13.1 (2021-05-27)
 
+* Fixes handling of relative imports for WDL specified using non-file URIs (e.g. http(s))
 * Fixes quoting and escaping of import URIs by code formatter/generator
 * Disables wrapping of expressions within placeholders